moved interface in a module
