+++ database/src/lib.rs
@@ -14,7 +14,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 //!
-//! The cuprate-db crates implement (as its name suggests) the relations between the blockchain/txpool objects and their databases.
+//! The cuprate-db crate implement (as its name suggests) the relations between the blockchain/txpool objects and their databases.
 //! `lib.rs` contains all the generics, trait and specification for interfaces between blockchain and a backend-agnostic database
 //! Every other files in this folder are implementation of these traits/methods to real storage engine.
 //! 
@@ -29,13 +29,9 @@
 #![deny(clippy::expect_used, clippy::panic)]
 #![allow(dead_code, unused_macros)] // temporary
 
-use database::{Database, Interface};
 use thiserror::Error;
-use monero::{Hash, Block, BlockHeader, consensus::Encodable, util::ringct::{RctSig, Key}, cryptonote::hash::keccak_256, TxOut};
-use transaction::{Cursor, Transaction,  WriteTransaction, DupCursor, DupWriteCursor};
-use types::{OutputMetadata, TransactionPruned};
-use core::slice::SlicePattern;
-use std::ops::{Range};
+use monero::{Hash, Block, BlockHeader, util::ringct::RctSig};
+use std::ops::Range;
 
 #[cfg(feature = "mdbx")]
 pub mod mdbx;
@@ -46,11 +42,11 @@ pub mod error;
 pub mod table;
 pub mod types;
 pub mod encoding;
+pub mod interface;
 
-const DEFAULT_BLOCKCHAIN_DATABASE_FILENAME: &str = "blockchain.db";
-const DEFAULT_TXPOOL_DATABASE_FILENAME: &str = "txpool_mem.db";
+const DEFAULT_BLOCKCHAIN_DATABASE_DIRECTORY: &str = "blockchain";
+const DEFAULT_TXPOOL_DATABASE_DIRECTORY: &str = "txpool_mem";
 const BINCODE_CONFIG: bincode::config::Configuration<bincode::config::LittleEndian, bincode::config::Fixint> = bincode::config::standard().with_fixed_int_encoding();
-const ZEROKVAL: [u8; 0] = [];
 
 // ------------------------------------------|      Database      |------------------------------------------
 
@@ -86,13 +82,11 @@ pub mod database {
     impl<'service,D: Database<'service>> Interface<'service,D> {
 
         fn from(db: Arc<D>) -> Result<Self,DB_FAILURES> {
-
 			Ok(Self { db, tx: None})
     	}	
 
 		fn open(&'service mut self) -> Result<(),DB_FAILURES> {
     		let tx = self.db.tx_mut().map_err(Into::into)?;
-
 			self.tx = Some(tx);
 			Ok(())
     	}
@@ -162,7 +156,7 @@ pub mod transaction {
 		
 		fn put_cursor_dup(&mut self, key: &T::Key, subkey: &T::SubKey, value: &T::Value) -> Result<(),DB_FAILURES>;
 
-		/// Delete all data under its key
+		/// Delete all data under associated to its key
 		fn del_nodup(&mut self) -> Result<(),DB_FAILURES>;
 	}
 
@@ -218,56 +212,11 @@ pub mod transaction {
 
 
 
-impl<'service, D: Database<'service>> Interface<'service,D> {
 
-	// --------------------------------| Blockchain |--------------------------------
-	
-	/// `height` fetch the current blockchain height.
-    ///
-    /// Return the current blockchain height. In case of failures, a DB_FAILURES will be return.
-    ///
-    /// No parameters is required.
-	fn height(&'service self) -> Result<Option<u64>,error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		let mut cursor = ro_tx.cursor::<table::blockhash>()?;
 
-		let last = cursor.last()?;
-		if let Some(pair) = last {
-			return Ok(Some(pair.1))
-		}
-		Ok(None)
-	}
 
-	/// `set_hard_fork_version` sets which hardfork version a height is on.<br>
-    ///
-    /// In case of failures, a `DB_FAILURES` will be return.
-    ///
-    /// Parameters:<br>
- 	/// `height`: is the height where the hard fork happen.<br>
-    /// `version`: is the version of the hard fork.
-    fn update_hard_fork_version(&'service self, height: u64, hardfork_version: u8) -> Result<(), error::DB_FAILURES> {
-		let b_hf = self.get::<table::blockhfversion>(&height)?;
 
-		if let Some(b_hf) = b_hf {
-			if b_hf != hardfork_version {
-				self.put::<table::blockhfversion>(&height, &hardfork_version)?;
-			}
-			Ok(())
-		} else {
-			Err(error::DB_FAILURES::DataNotFound)
-		}
-	}
 
-	/// `get_hard_fork_version` checks which hardfork version a height is on.
-    ///
-    /// In case of failures, a `DB_FAILURES` will be return.
-    ///
-    /// Parameters:<br>
-    /// `height`: is the height to check.
-    fn get_hard_fork_version(&'service self, height: u64) -> Result<Option<u8>, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		ro_tx.get::<table::blockhfversion>(&height)
-	}
 
 
 
@@ -279,373 +228,8 @@ impl<'service, D: Database<'service>> Interface<'service,D> {
 
 
 
-
-
-
 
-	// --------------------------------| Blocks |--------------------------------
-
-	/// `pop_block` pops the top block off the blockchain. This cause the last block to be deleted
-	/// from all its reference tables, including transactions it was refering to.
-    ///
-	/// Return the block that was popped. In case of failures, a `DB_FAILURES` will be return.
-	///
-    /// No parameters is required.
-    fn pop_block(&'service self) -> Result<Option<Block>, error::DB_FAILURES> {
-		let current_height = self.height()?;
-
-		if let Some(current_height) = current_height{
-			let blk = self.get::<table::blocks>(&current_height)?;
-
-			// Get the block and delete it from block's tables
-			if let Some(blk) = blk {
-				self.delete::<table::blocks>(&current_height, &None)?;
-				self.delete::<table::blockmetadata>(&current_height, &None)?;
-
-				// Re-encoding the slice and get the hash
-				let e_blk = bincode::encode_to_vec(&blk, BINCODE_CONFIG)
-					.map_err(|e| error::DB_FAILURES::SerializeIssue(error::DB_SERIAL::BincodeEncode(e)))?;
-				let hash = Hash::new(keccak_256(e_blk.as_slice())).into();
-				
-				self.delete::<table::blockhash>(&hash, &None)?;
-
-				// Now let's delete all its transactions
-				blk.0.tx_hashes.iter().for_each(|tx| {
-
-					todo!()
-				});
-				return Ok(Some(blk.0))
-			}
-		}
-		Ok(None)
-	}
-
-
-
-
-
-
-
-
-
-
-	// ------------------------------|  Transactions  |-----------------------------
-
-	/// `get_num_tx` fetches the total number of transactions stored in the database
-    ///
-    /// Should return the count. In case of failure, a DB_FAILURES will be return.
-    ///
-    /// No parameters is required.
-    fn get_num_tx(&'service self) -> Result<u64, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		ro_tx.num_entries::<table::txsprefix>().map(|n| n as u64)
-	}
-
-	/// `tx_exists` check if a transaction exist with the given hash.
-    ///
-    /// Return `true` if the transaction exist, `false` otherwise. In case of failure, a DB_FAILURES will be return.
-    ///
-    /// Parameters :
-    /// `h` is the given hash of transaction to check.
-    ///  `tx_id` is an optional mutable reference to get the transaction id out of the found transaction.
-    fn tx_exists(&'service self, hash: Hash) -> Result<bool, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		Ok(ro_tx.get::<table::txsidentifier>(&hash.into())?.is_some())
-	}
-
-	/// `get_tx_unlock_time` fetch a transaction's unlock time/height
-    ///
-    /// Should return the unlock time/height in u64. In case of failure, a DB_FAILURES will be return.
-    ///
-    /// Parameters:
-    /// `h`: is the given hash of the transaction to check.
-    fn get_tx_unlock_time(&'service self, hash: Hash) -> Result<u64, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		let txindex = ro_tx.get::<table::txsidentifier>(&hash.into())?
-			.ok_or(error::DB_FAILURES::NoneFound("wasn't able to find a transaction in the database"))?;
-		Ok(txindex.unlock_time)
-	}
-
-	/// `get_tx` fetches the transaction with the given hash.
-    ///
-    /// Should return the transaction. In case of failure, a DB_FAILURES will be return.
-    ///
-    /// Parameters:
-    /// `h`: is the given hash of transaction to fetch.
-    fn get_tx(&'service self, hash: Hash) -> Result<Option<monero::Transaction>, error::DB_FAILURES> {
-		let pruned_tx = self.get_pruned_tx(hash)?
-			.ok_or(error::DB_FAILURES::NoneFound("failed to find prefix of a transaction"))?;
-		
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		let txindex = ro_tx.get::<table::txsidentifier>(&hash.into())?
-			.ok_or(error::DB_FAILURES::NoneFound("failed to find index of a transaction"))?;
-		let prunable_part = ro_tx.get::<table::txsprunable>(&txindex.tx_id)?
-			.ok_or(error::DB_FAILURES::NoneFound("failed to find prunable part of a transaction"))?;
-
-		Ok(Some(pruned_tx.to_transaction(prunable_part.as_slice())
-			.map_err(|e| error::DB_FAILURES::SerializeIssue(error::DB_SERIAL::ConsensusDecode(prunable_part)))?))
-	}
-
-	/// `get_pruned_tx` fetches the transaction base with the given hash.
-    ///
-    /// Should return the transaction. In case of failure, a DB_FAILURES will be return.
-    ///
-    /// Parameters:
-    /// `h`: is the given hash of transaction to fetch.
-    fn get_pruned_tx(&'service self, hash: Hash) -> Result<Option<TransactionPruned>, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-
-		let txindex = ro_tx.get::<table::txsidentifier>(&hash.into())?
-			.ok_or(error::DB_FAILURES::NoneFound("wasn't able to find a transaction in the database"))?;
-		ro_tx.get::<table::txsprefix>(&txindex.tx_id)
-	}
-
-
-
-
-
-
-	// --------------------------------|  Outputs  |--------------------------------
-
-	/// `add_output` add an output data to it's storage .
-    ///
-    /// It internally keep track of the global output count. The global output count is also used to index outputs based on
-    /// their order of creations.
-    ///
-    /// Should return the amount output index. In case of failures, a DB_FAILURES will be return.
-	///
-    /// Parameters:
-    /// `tx_hash`: is the hash of the transaction where the output comes from.
-    /// `output`: is the output's publickey to store.
-	/// `index`: is the local output's index (from transaction).
-	/// `unlock_time`: is the unlock time (height) of the output.
-	/// `commitment`: is the RingCT commitment of this output.
-    fn add_output(&'service self,
-		tx_hash: Hash,
-		output: TxOut,
-		local_index: u64,
-		unlock_time: u64,
-		commitment: Option<Key>,
-    	) -> Result<u64, error::DB_FAILURES> {
-
-		let height = self.height()?
-			.ok_or(error::DB_FAILURES::NoneFound("add_output() didn't find a blockchain height"))?;
-
-		let pubkey = output.target.as_one_time_key().map(Into::into);
-
-		// RingCT Outputs
-		if let Some(commitment) = commitment {
-
-			let amount_index = self.get_rct_num_outputs()?;
-
-			let out_metadata = OutputMetadata {
-				tx_hash: tx_hash.into(),
-				local_index,
-				pubkey,
-				unlock_time,
-				height,
-				commitment: Some(commitment.into()),
-			};
-
-			self.put::<table::outputmetadata>(&amount_index, &out_metadata)?;
-			Ok(amount_index)
-		}
-		// Pre-RingCT Outputs
-		else {
-
-			let amount_index = self.get_pre_rct_num_outputs(output.amount.0)? + 1;
-
-			let out_metadata = OutputMetadata {
-				tx_hash: tx_hash.into(),
-				local_index,
-				pubkey,
-				unlock_time,
-				height,
-				commitment: None,
-			};
-
-			let mut cursor = self.write_cursor_dup::<table::prerctoutputmetadata>()?;
-			cursor.put_cursor_dup(&output.amount.0, &amount_index, &out_metadata)?;
-			Ok(amount_index)
-		}
-	}
-
-	/// `get_output` get an output's data
-    ///
-    /// Return the public key, unlock time, and block height for the output with the given amount and index, collected in a struct
-    /// In case of failures, a `DB_FAILURES` will be return. Precisely, if the output cannot be found, an `OUTPUT_DNE` error will be return.
-    /// If any of the required part for the final struct isn't found, a `DB_ERROR` will be return
-    ///
-    /// Parameters:
-    /// `amount`: is the corresponding amount of the output
-    /// `index`: is the output's index (indexed by amount)
-    /// `include_commitment` : `true` by default.
-    fn get_output(&'service self, amount: Option<u64>, index: u64) -> Result<Option<OutputMetadata>, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		if let Some(amount) = amount {
-			let mut cursor = ro_tx.cursor_dup::<table::prerctoutputmetadata>()?;
-			cursor.get_dup(&amount, &index)
-		} else {
-			ro_tx.get::<table::outputmetadata>(&index)
-		}
-	}
-
-    /// `get_output_list` gets a collection of output's data from a corresponding index collection.
-    ///
-    /// Return a collection of output's data. In case of failurse, a `DB_FAILURES` will be return.
-    ///
-    /// Parameters:
-    /// `amounts`: is the collection of amounts corresponding to the requested outputs.
-    /// `offsets`: is a collection of outputs' index (indexed by amount).
-    /// `allow partial`: `false` by default.
-    fn get_output_list(
-		&'service self,
-		amounts: Option<Vec<u64>>,
-		offsets: Vec<u64>,
-    	) -> Result<Option<Vec<OutputMetadata>>, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		let mut result: Vec<OutputMetadata> = Vec::new();
-		
-		// Pre-RingCT output to be found.
-		if let Some(amounts) = amounts {
-			let mut cursor = ro_tx.cursor_dup::<table::prerctoutputmetadata>()?;
-
-			for ofs in amounts.into_iter().zip(offsets) {
-
-				if ofs.0 == 0 {
-					let output = ro_tx.get::<table::outputmetadata>(&ofs.1)?
-						.ok_or(error::DB_FAILURES::NoneFound("An output hasn't been found in the database"))?;
-					result.push(output);
-				} else {
-					let output = cursor.get_dup(&ofs.0, &ofs.1)?
-						.ok_or(error::DB_FAILURES::NoneFound("An output hasn't been found in the database"))?;
-					result.push(output);
-				}
-			}
-		// No Pre-RingCT outputs to be found.
-		} else {
-			for ofs in offsets {
-
-				let output = ro_tx.get::<table::outputmetadata>(&ofs)?
-					.ok_or(error::DB_FAILURES::NoneFound("An output hasn't been found in the database"))?;
-				result.push(output);				
-			}
-		}
-
-		Ok(Some(result))
-	}
-
-    /// `get_num_outputs` fetches the number post-RingCT output.
-    ///
-    /// Return the number of post-RingCT outputs. In case of failures a `DB_FAILURES` will be return.
-    ///
-    /// Parameters:
-    /// `amount`: is the output amount being looked up.
-    fn get_rct_num_outputs(&'service self) -> Result<u64, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		
-		ro_tx.num_entries::<table::outputmetadata>().map(|n| n as u64)
-	}
-
-	/// `get_pre_rct_num_outputs` fetches the number of preRCT outputs of a given amount.
-    ///
-    /// Return a count of outputs of the given amount. in case of failures a `DB_FAILURES` will be return.
-    ///
-    /// Parameters:
-    /// `amount`: is the output amount being looked up.
-	fn get_pre_rct_num_outputs(&'service self, amount: u64) -> Result<u64, error::DB_FAILURES> {
-		let ro_tx = self.db.tx().map_err(Into::into)?;
-		let mut cursor = ro_tx.cursor_dup::<table::prerctoutputmetadata>()?;
-
-		transaction::Cursor::set(&mut cursor, &amount)?;
-		let out_metadata: Option<(u64, OutputMetadata)> = transaction::DupCursor::last_dup(&mut cursor)?;
-		if let Some(out_metadata) = out_metadata {
-			return Ok(out_metadata.0)
-		}
-		Err(error::DB_FAILURES::Other("failed to decode the subkey and value"))
-	}
-}
-
-
-
-
-
-
-
-
-
-
-
-/*
-Errors not yet implemented: 
-- MapFull
-- VersionMismatch
-- Invalid
-- PageFull
-- UnableExtendMapsize
-- Incompatible
-- DbsFull - Useless since we used only two database
-- BadTxn
-- BadValSize
-- BadDbi
-- Problem
-- Busy
-- Multival
-- WannaRecovery
-- KeyMismtach
-- InvalidValue
-- Access
-- DecodeError
-*/
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+/////////////////////// MONEROD BREAKOUT ////////////////////////////
 
 
 
@@ -785,11 +369,7 @@ pub enum DB_FAILURES {
     HASH_DNE(Option<Hash>),
 }
 
-pub trait KeyValueDatabase {
-    fn add_data_to_cf<D: ?Sized + Encodable>(cf: &str, data: &D) -> Result<Hash, DB_FAILURES>;
-}
-
-pub trait BlockchainDB: KeyValueDatabase {
+pub trait BlockchainDBaaaa {
     // supposed to be private
 
     // TODO: understand
